Move all classes out of types module
